requirements.txt
cloudpickle==3.1.1
Farama-Notifications==0.0.4
filelock==3.18.0
fsspec==2025.5.1
gymnasium==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
typing_extensions==4.14.0
tzdata==2025.2
